Move SimulationConfig to rumoca-session::config
Shared TOML configuration model per the naming-scheme proposal. Aggregates
input + signal sections from rumoca-input and transport config from
rumoca-transport-*. Consumed by sim-fb today; CLI and wasm bindings will
consume it directly after sim-fb is dissolved.

- crates/rumoca-session/src/config.rs: new module, SimFbConfig renamed
  to SimulationConfig
- rumoca-session/Cargo.toml: adds rumoca-codec-flatbuffers, rumoca-input,
  rumoca-transport-udp deps for re-exports
- rumoca-sim-fb/src/config.rs: now just re-exports from rumoca-session;
  transitional `pub type SimFbConfig = SimulationConfig` alias keeps
  existing sim_loop.rs imports working

Tests covering parse (rover standalone + quadrotor + partial-FB rejection)
moved to rumoca-session. Workspace clippy clean.
